move TeamCalendarFragment to under team
