+++ app/src/main/java/org/ole/planet/myplanet/ui/community/CommunityPagerAdapter.kt
@@ -6,7 +6,7 @@ import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import org.ole.planet.myplanet.MainApplication.Companion.context
 import org.ole.planet.myplanet.R
-import org.ole.planet.myplanet.ui.enterprises.TeamCalendarFragment
+import org.ole.planet.myplanet.ui.team.TeamCalendarFragment
 import org.ole.planet.myplanet.ui.enterprises.FinanceFragment
 import org.ole.planet.myplanet.ui.enterprises.ReportsFragment
 import org.ole.planet.myplanet.ui.news.NewsFragment
