+++ app/src/main/res/layout/fragment_enterprise_calendar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/secondary_bg"
     android:layout_height="match_parent"
-    tools:context=".ui.enterprises.TeamCalendarFragment">
+    tools:context=".ui.team.TeamCalendarFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
